+++ src/Model/Helper.cpp
@@ -33,3 +33,10 @@ QMap<int, QString> Helper::getAllGroups()
 
 	return groups;
 }
+
+void Helper::delay( int secs )
+{
+    QTime dieTime= QTime::currentTime().addSecs( secs );
+    while (QTime::currentTime() < dieTime)
+        QCoreApplication::processEvents( QEventLoop::AllEvents, 100 );
+}